+++ xarray/tests/test_cftimeindex.py
@@ -586,6 +586,14 @@ def test_indexing_in_series_iloc(series, index):
 
 
 @pytest.mark.skipif(not has_cftime, reason='cftime not installed')
+def test_series_dropna(index):
+    series = pd.Series([0., 1., np.nan, np.nan], index=index)
+    expected = series.iloc[:2]
+    result = series.dropna()
+    assert result.equals(expected)
+
+
+@pytest.mark.skipif(not has_cftime, reason='cftime not installed')
 def test_indexing_in_dataframe_loc(df, index, scalar_args, range_args):
     expected = pd.Series([1], name=index[0])
     for arg in scalar_args: